+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Bulk process bundle items, reported via issue [#1](https://github.com/netresearch/module-shipping-inventory/issues/1)
+
 ## 1.1.0
 
 Magento 2.4.4 compatibility release
@@ -25,5 +31,3 @@ Magento 2.4.4 compatibility release
 ## 1.0.0
 
 Initial release 
-
-